server/requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# HTTP客户端
httpx==0.25.2
aiohttp==3.9.1

# 环境变量
python-dotenv==1.0.0

# 数据验证
pydantic==2.5.0
email-validator==2.1.0

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 数据库（可选）
pymongo==4.6.0
motor==3.3.2

# 缓存
redis==5.0.1
aioredis==2.0.1

# AI和翻译服务
openai==1.3.8
azure-ai-translation-text==1.0.0
google-cloud-translate==3.12.1
deepl==1.16.1

# 本地模型支持（可选）
# 如果需要使用本地模型，请取消注释以下依赖：
# torch>=2.0.0
# transformers>=4.30.0
# sentencepiece>=0.1.99
# sacremoses>=0.0.53

# 工具库
python-dateutil==2.8.2
pytz==2023.3
psutil==5.9.6

# 开发和测试
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0 